+++ functions/__vmware_workstation__.py
@@ -4,21 +4,17 @@ import subprocess
 import shutil
 import logging
 
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 class VMwareInstaller:
     PKGVER = "17.5.2"
     BUILDVER = "23775571"
     TOOLS_VERSION = "12.4.0-23259341"
-    DKMS_MODULE = "vmware-host-modules"
     CARCH = "x86_64"
     BASE_URL = f"https://softwareupdate.vmware.com/cds/vmw-desktop/ws/{PKGVER}/{BUILDVER}/linux/packages"
     BUNDLE_URL = f"https://softwareupdate.vmware.com/cds/vmw-desktop/ws/{PKGVER}/{BUILDVER}/linux/core/VMware-Workstation-{PKGVER}-{BUILDVER}.{CARCH}.bundle.tar"
     BUNDLE_FILENAME = f"VMware-Workstation-{PKGVER}-{BUILDVER}.{CARCH}.bundle.tar"
     CACHE_DIR = Path(os.path.expanduser("~")) / ".cache" / "vcandy" / "VMware"
     CACHE_DIR.mkdir(parents=True, exist_ok=True)
-    DEPENDENCIES = []
-    EXTRACTED_DIR = os.path.join(CACHE_DIR, "extracted_components")
     COMPONENT_FILENAMES = [
         f"vmware-tools-linux-{TOOLS_VERSION}.{CARCH}.component.tar",
         f"vmware-tools-linuxPreGlibc25-{TOOLS_VERSION}.{CARCH}.component.tar",
@@ -26,14 +22,19 @@ class VMwareInstaller:
         f"vmware-tools-solaris-{TOOLS_VERSION}.{CARCH}.component.tar",
         f"vmware-tools-windows-{TOOLS_VERSION}.{CARCH}.component.tar",
         f"vmware-tools-winPre2k-{TOOLS_VERSION}.{CARCH}.component.tar",
-        f"vmware-tools-winPreVista-{TOOLS_VERSION}.{CARCH}.component.tar",
+        f"vmware-tools-winPreVista-{TOOLS_VERSION}.{CARCH}.component.tar"
     ]
+    EXTRACTED_DIR = os.path.join(CACHE_DIR, "extracted_components")
+    DEPENDENCIES = []
+    PACKAGE_NAME = "vmware-host-modules"
+    PACKAGE_VERSION = PKGVER # Do not make the PACKAGE_VERSION to PKGVER
 
     def __init__(self, hide, action, linux_distro):
+        self.COMPONENT_URLS = [f"{self.BASE_URL}/{filename}" for filename in self.COMPONENT_FILENAMES]
+        self.setup_logging()
         self.hide = hide
         self.action = action
         self.linux_distro = linux_distro
-        self.COMPONENT_URLS = [f"{self.BASE_URL}/{filename}" for filename in self.COMPONENT_FILENAMES]
 
         if self.linux_distro == "fedora":
             self.DEPENDENCIES = [
@@ -61,16 +62,22 @@ class VMwareInstaller:
         elif self.action == "remove":
             self.uninstall_vmware()
 
-    @staticmethod
-    def run_command(command):
+    def setup_logging(self):
+        """Setup logging configuration."""
+        logging.basicConfig(level=logging.INFO,
+                            format='%(asctime)s - %(levelname)s - %(message)s',
+                            handlers=[logging.FileHandler("vmware_installer.log"),
+                                      logging.StreamHandler()])
+
+    def run_command(self, command):
         """Run a shell command and print its output."""
-        try:
-            result = subprocess.run(command, shell=True, text=True, capture_output=True)
-            result.check_returncode()
-            logging.info(result.stdout)
-        except subprocess.CalledProcessError as e:
-            logging.error(f"Command failed: {command}")
-            logging.error(e.stderr)
+        logging.info(f"Running command: {command}")
+        result = subprocess.run(command, shell=True, text=True, capture_output=True)
+        if result.returncode != 0:
+            logging.error(f"Command failed: {command}\n{result.stderr}")
+            return False
+        logging.info(result.stdout)
+        return True
 
     def download_file(self, url, filename):
         """Download a file from a URL using wget."""
@@ -101,15 +108,18 @@ class VMwareInstaller:
         
         logging.info(f"Folder {folder_to_clone} has been cloned to {clone_location}/{folder_to_clone}")
 
-
     def install_vmware_modules(self):
-        logging.info("Cloning vmware-host-modules repository...")
-        self.run_command(f"git clone -b tmp/workstation-17.5.2-k6.9.1 https://github.com/nan0desu/vmware-host-modules.git {self.CACHE_DIR}/vmware-host-modules")
+        """Install VMware modules."""
+        logging.info(f"Cloning {self.PACKAGE_NAME} repository...")
+        self.run_command(f"git clone -b tmp/workstation-17.5.0-k6.8 https://github.com/mkubecek/{self.PACKAGE_NAME}.git")
 
-        os.chdir(f"{self.CACHE_DIR}/vmware-host-modules")
+        logging.info("Getting the DKMS modules")
+        self.sparse_checkout("https://github.com/Hakanbaban53/Container-and-Virtualization-Installer", "Adding_the_vmware_worksitation_support_fedora", "vmware_dkms_files", f"{self.CACHE_DIR}/vmware_dkms_files")
+
+        os.chdir(f"{self.CACHE_DIR}/vmware_dkms_files")
         logging.info("Making and copying vmmon.tar and vmnet.tar...")
         self.run_command("make tarballs")
-        self.run_command("sudo cp -v vmmon.tar vmnet.tar /usr/lib/vmware/modules/source/")
+        self.run_command(f"sudo cp -v vmmon.tar vmnet.tar /usr/lib/vmware/modules/source/")
 
         logging.info("Extracting vmmon.tar and vmnet.tar...")
         os.chdir("/usr/lib/vmware/modules/source/")
@@ -117,46 +127,39 @@ class VMwareInstaller:
         self.run_command("sudo tar -xvf vmnet.tar")
 
         logging.info("Creating directories for DKMS...")
-        self.run_command(f"sudo mkdir -p /usr/src/{self.DKMS_MODULE}-{self.PKGVER}")
+        self.run_command(f"sudo mkdir -p /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}")
 
         logging.info("Moving source files to DKMS directory...")
-        self.run_command(f"sudo cp -r vmmon-only /usr/src/{self.DKMS_MODULE}-{self.PKGVER}/")
-        self.run_command(f"sudo cp -r vmnet-only /usr/src/{self.DKMS_MODULE}-{self.PKGVER}/")
+        self.run_command(f"sudo cp -r vmmon-only /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/")
+        self.run_command(f"sudo cp -r vmnet-only /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/")
 
-        logging.info("Getting the DKMS modules")
-        self.sparse_checkout("https://github.com/Hakanbaban53/Container-and-Virtualization-Installer", "Adding_the_vmware_worksitation_support_fedora", "vmware_dkms_files", f"{self.CACHE_DIR}/vmware_dkms_files")
-
-        logging.info("Copying Makefile to DKMS directory...")
-        self.run_command(f"sudo cp -r {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/Makefile /usr/src/{self.DKMS_MODULE}-{self.PKGVER}/")
+        logging.info("Copying Makefile and dkms.conf to DKMS directory...")
+        self.run_command(f"sudo cp -r {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/Makefile /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/")
 
         logging.info("Creating DKMS configuration for vmware-host-modules...")
-        dkms_conf_vmware_host_modules = """
-PACKAGE_NAME="vmware-host-modules"
-PACKAGE_VERSION="17.5.2"
-MAKE="make KVERSION=$kernelver SRCDIR=/usr/src/$PACKAGE_NAME-$PACKAGE_VERSION"
-CLEAN="make clean"
-AUTOINSTALL="YES"
+        with open(f"{self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/dkms.conf", "r") as template_file:
+            dkms_conf_template = template_file.read()
 
-BUILT_MODULE_NAME[0]="vmmon"
-BUILT_MODULE_LOCATION[0]='vmmon-only'
-DEST_MODULE_LOCATION[0]="/kernel/drivers/misc"
-
-BUILT_MODULE_NAME[1]="vmnet"
-BUILT_MODULE_LOCATION[1]='vmnet-only'
-DEST_MODULE_LOCATION[1]="/kernel/drivers/net"
-        """
-        with open("/tmp/vmware-host-modules-17.5.2-dkms.conf", "w") as f:
-            f.write(dkms_conf_vmware_host_modules)
-        self.run_command("sudo mv /tmp/vmware-host-modules-17.5.2-dkms.conf /usr/src/vmware-host-modules-17.5.2/dkms.conf")
+        dkms_conf_vmware_host_modules = dkms_conf_template.format(
+            PACKAGE_NAME=self.PACKAGE_NAME,
+            PACKAGE_VERSION=self.PACKAGE_VERSION
+        )
+        
+        temp_conf_path = f"/tmp/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}-dkms.conf"
+        with open(temp_conf_path, "w") as conf_file:
+            conf_file.write(dkms_conf_vmware_host_modules)
+        
+        self.run_command(f"sudo mv {temp_conf_path} /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/dkms.conf")
+        logging.info(f"DKMS configuration file created at /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/dkms.conf")
 
         logging.info("Applying patches...")
-        self.run_command(f"sudo patch -p2 -d /usr/src/{self.DKMS_MODULE}-{self.PKGVER}/vmmon-only < {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/vmmon.patch --force")
-        self.run_command(f"sudo patch -p2 -d /usr/src/{self.DKMS_MODULE}-{self.PKGVER}/vmnet-only < {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/vmnet.patch --force")
+        self.run_command(f"sudo patch -p2 -d /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/vmmon-only < {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/vmmon.patch")
+        self.run_command(f"sudo patch -p2 -d /usr/src/{self.PACKAGE_NAME}-{self.PACKAGE_VERSION}/vmnet-only < {self.CACHE_DIR}/vmware_dkms_files/vmware_dkms_files/vmnet.patch")
 
         logging.info("Adding and building vmware-host-modules module with DKMS...")
-        self.run_command(f"sudo dkms add -m {self.DKMS_MODULE} -v {self.PKGVER}")
-        self.run_command(f"sudo dkms build -m {self.DKMS_MODULE} -v {self.PKGVER}")
-        self.run_command(f"sudo dkms install -m {self.DKMS_MODULE} -v {self.PKGVER}")
+        self.run_command(f"sudo dkms add -m {self.PACKAGE_NAME} -v {self.PACKAGE_VERSION}")
+        self.run_command(f"sudo dkms build -m {self.PACKAGE_NAME} -v {self.PACKAGE_VERSION}")
+        self.run_command(f"sudo dkms install -m {self.PACKAGE_NAME} -v {self.PACKAGE_VERSION}")
 
         logging.info("Running vmware-modconfig to install all modules...")
         self.run_command("sudo vmware-modconfig --console --install-all")
@@ -164,6 +167,7 @@ DEST_MODULE_LOCATION[1]="/kernel/drivers/net"
         os.chdir("..")  # Return to previous directory
 
     def create_service_files(self):
+        """Create systemd service files for VMware."""
         services = {
             'vmware-networks-configuration.service': """
 [Unit]
@@ -205,14 +209,12 @@ WantedBy=multi-user.target
             self.run_command(f"sudo mv /tmp/{filename} /etc/systemd/system/{filename}")
 
     def install_dependencies(self):
+        """Install necessary dependencies for VMware."""
         logging.info("Installing dependencies...")
-        
-        if self.linux_distro == "fedora":
-            self.run_command(f"sudo dnf install -y {' '.join(self.DEPENDENCIES)}")
-        elif self.linux_distro in {"debian", "ubuntu"}:
-            self.run_command(f"sudo apt install -y {' '.join(self.DEPENDENCIES)}")
+        self.run_command(f"sudo dnf install -y {' '.join(self.DEPENDENCIES)}")
 
     def enable_services(self):
+        """Enable and start VMware services."""
         logging.info("Enabling and starting VMware services...")
         services = [
             "vmware-networks-configuration.service",
@@ -223,11 +225,13 @@ WantedBy=multi-user.target
             self.run_command(f"sudo systemctl enable --now {service}")
 
     def add_user_to_vmware_group(self):
+        """Add the current user to the vmware group."""
         user = os.getlogin()
         logging.info(f"Adding {user} to the vmware group...")
         self.run_command(f"sudo usermod -aG vmware {user}")
 
     def install_vmware(self):
+        """Perform the full VMware installation."""
         logging.info("Step 1: Downloading the VMware Workstation installer and components...")
         os.makedirs(self.CACHE_DIR, exist_ok=True)
         self.download_file(self.BUNDLE_URL, self.BUNDLE_FILENAME)
@@ -252,6 +256,7 @@ WantedBy=multi-user.target
         )
         self.run_command(install_command)
 
+
         logging.info("Step 5: Installing necessary dependencies...")
         self.install_dependencies()
 
@@ -270,9 +275,10 @@ WantedBy=multi-user.target
         logging.info(f"VMware installation and setup on {self.linux_distro} is complete.")
 
     def uninstall_vmware(self):
+        """Uninstall VMware Workstation."""
         logging.info("Uninstalling VMware Workstation...")
 
-        logging.info("Stopping and disabling VMware services...")
+        logging.info("Step 1: Stopping and disabling VMware services...")
         services = [
             "vmware-networks-configuration.service",
             "vmware-usbarbitrator.path",
@@ -282,25 +288,30 @@ WantedBy=multi-user.target
             self.run_command(f"sudo systemctl stop {service}")
             self.run_command(f"sudo systemctl disable {service}")
 
-        logging.info("Removing systemd service files...")
+        logging.info("Step 2: Removing systemd service files...")
         for service in services:
             service_file = f"/etc/systemd/system/{service}"
             if os.path.exists(service_file):
                 self.run_command(f"sudo rm {service_file}")
 
-        logging.info("Removing VMware modules from DKMS...")
-        self.run_command(f"sudo dkms remove -m {self.DKMS_MODULE} -v {self.PKGVER} --all")
+        logging.info("Step 3: Removing VMware modules from DKMS...")
+        self.run_command(f"sudo dkms remove -m {self.PACKAGE_NAME} -v {self.PACKAGE_VERSION} --all")
 
-        logging.info("Running the uninstallation script...")
+
+        logging.info("Step 4: Removing VMware modules...")
+        self.run_command("sudo rm -rf /usr/lib/vmware")
+
+        logging.info("Step 5: Running the uninstallation script...")
         uninstall_script = "/usr/bin/vmware-installer"
         if os.path.exists(uninstall_script):
             self.run_command(f'echo "yes" | sudo {uninstall_script} --uninstall-product vmware-workstation')
 
-        logging.info("Removing extracted components directory...")
+        logging.info("Step 6: Removing extracted components directory...")
         if os.path.exists(self.EXTRACTED_DIR):
             shutil.rmtree(self.EXTRACTED_DIR)
 
         logging.info(f"VMware uninstallation on {self.linux_distro} is complete.")
+
 
 if __name__ == "__main__":
     VMwareInstaller(False, "install", "fedora")
